Final Version,

+++ AlliedSchool/AlliedSchool/Item.cs
@@ -21,8 +21,8 @@ namespace AlliedSchool
     
         public int ItemId { get; set; }
         public string Name { get; set; }
-        public long Price { get; set; }
-        public long Quantity { get; set; }
+        public int Price { get; set; }
+        public int Quantity { get; set; }
     
         public virtual ICollection<ShoppingItem> ShoppingItems { get; set; }
     }
